Locally efficient doubly robust DiD estimator for the ATT, with panel data

drdid_panel is used to compute the locally efficient doubly robust estimators for the ATT in difference-in-differences (DiD) setups with panel data.

drdid_panel(
  y1,
  y0,
  D,
  covariates,
  i.weights = NULL,
  boot = FALSE,
  boot.type = "weighted",
  nboot = NULL,
  inffunc = FALSE
)

Arguments

y1
An \(n\) x \(1\) vector of outcomes from the post-treatment period.

y0
An \(n\) x \(1\) vector of outcomes from the pre-treatment period.

D
An \(n\) x \(1\) vector of Group indicators (=1 if observation is treated in the post-treatment, =0 otherwise).

covariates
An \(n\) x \(k\) matrix of covariates to be used in the propensity score and regression estimation. If covariates = NULL, this leads to an unconditional DID estimator.

i.weights
An \(n\) x \(1\) vector of weights to be used. If NULL, then every observation has the same weights.

boot
Logical argument to whether bootstrap should be used for inference. Default is FALSE.

boot.type
Type of bootstrap to be performed (not relevant if boot = FALSE). Options are "weighted" and "multiplier". If boot = TRUE, default is "weighted".

nboot
Number of bootstrap repetitions (not relevant if boot = FALSE). Default is 999.

inffunc
Logical argument to whether influence function should be returned. Default is FALSE.

Value

A list containing the following components:

ATT

The DR DID point estimate.

se

The DR DID standard error.

uci

Estimate of the upper bound of a 95% CI for the ATT.

lci

Estimate of the lower bound of a 95% CI for the ATT.

boots

All Bootstrap draws of the ATT, in case bootstrap was used to conduct inference. Default is NULL.

att.inf.func

Estimate of the influence function. Default is NULL.

call.param

The matched call.

argu

Some arguments used (explicitly or not) in the call (panel = TRUE, estMethod = "trad", boot, boot.type, nboot, type="dr")

Details

The drdid_panel function implements the locally efficient doubly robust difference-in-differences (DID) estimator for the average treatment effect on the treated (ATT) defined in equation (3.1) in Sant'Anna and Zhao (2020). This estimator makes use of a logistic propensity score model for the probability of being in the treated group, and of a linear regression model for the outcome evolution among the comparison units.

The propensity score parameters are estimated using maximum likelihood, and the outcome regression coefficients are estimated using ordinary least squares.

References

Sant'Anna, Pedro H. C. and Zhao, Jun. (2020), "Doubly Robust Difference-in-Differences Estimators." Journal of Econometrics, Vol. 219 (1), pp. 101-122, doi: 10.1016/j.jeconom.2020.06.003

Examples

# Form the Lalonde sample with CPS comparison group (data in wide format)
eval_lalonde_cps <- subset(nsw, nsw$treated == 0 | nsw$sample == 2)
# Select some covariates
covX = as.matrix(cbind(eval_lalonde_cps$age, eval_lalonde_cps$educ,
                       eval_lalonde_cps$black, eval_lalonde_cps$married,
                       eval_lalonde_cps$nodegree, eval_lalonde_cps$hisp,
                       eval_lalonde_cps$re74))

# Implement traditional DR locally efficient DID with panel data
drdid_panel(y1 = eval_lalonde_cps$re78, y0 = eval_lalonde_cps$re75,
             D = eval_lalonde_cps$experimental,
             covariates = covX)
#>  Call:
#> drdid_panel(y1 = eval_lalonde_cps$re78, y0 = eval_lalonde_cps$re75, 
#>     D = eval_lalonde_cps$experimental, covariates = covX)
#> ------------------------------------------------------------------
#>  Locally efficient DR DID estimator for the ATT:
#>  
#>    ATT     Std. Error  t value    Pr(>|t|)  [95% Conf. Interval] 
#> -871.3271   396.0332   -2.2001     0.0278   -1647.5521  -95.1022 
#> ------------------------------------------------------------------
#>  Estimator based on panel data.
#>  Outcome regression est. method: OLS.
#>  Propensity score est. method: maximum likelihood.
#>  Analytical standard error.
#> ------------------------------------------------------------------
#>  See Sant'Anna and Zhao (2020) for details.