Inverse Probability Weighted Difference-in-Differences Estimator, with Repeated Cross Section Data

Inverse Probability Weighted Difference-in-Differences Estimator, with Repeated Cross Section Data

ipw_did_rc(
  y,
  post,
  D,
  covariates,
  i.weights = NULL,
  boot = F,
  boot.type = "weighted",
  nboot = NULL,
  inffunc = F
)

Arguments

y
An \(n\) x \(1\) vector of outcomes from the both pre and post-treatment periods.

post
An \(n\) x \(1\) vector of Post-Treatment dummies (post = 1 if observation belongs to post-treatment period, and post = 0 if observation belongs to pre-treatment period.)

D
An \(n\) x \(1\) vector of Group indicators (=1 if observation is treated in the post-treatment, =0 otherwise).

covariates
An \(n\) x \(k\) matrix of covariates to be used in the propensity score estimation. If covariates = NULL, this leads to an unconditional DID estimator.

i.weights
An \(n\) x \(1\) vector of weights to be used. If NULL, then every observation has the same weights.

boot
Logical argument to whether bootstrap should be used for inference. Default is FALSE.

boot.type
Type of bootstrap to be performed (not relevant if boot = FALSE). Options are "weighted" and "multiplier". If boot==T, default is "weighted".

nboot
Number of bootstrap repetitions (not relevant if boot = FALSE). Default is 999 if boot = TRUE.

inffunc
Logical argument to whether influence function should be returned. Default is FALSE.

Value

A list containing the following components:

ATT

The IPW DID point estimate.

se

The IPW DID standard error

uci

Estimate of the upper bound of a 95% CI for the ATT

lci

Estimate of the lower bound of a 95% CI for the ATT

boots

All Bootstrap draws of the ATT, in case bootstrap was used to conduct inference. Default is NULL

att.inf.func

Estimate of the influence function. Default is NULL

call.param

The matched call.

argu

Some arguments used (explicitly or not) in the call (panel = F, normalized = F, boot, boot.type, nboot, type="ipw")

References

Abadie, Alberto (2005), "Semiparametric Difference-in-Differences Estimators", Review of Economic Studies, vol. 72(1), p. 1-19, https://doi.org/10.1111/0034-6527.00321. Sant'Anna, Pedro H. C. and Zhao, Jun. (2020), "Doubly Robust Difference-in-Differences Estimators." Journal of Econometrics, Forthcoming, https://arxiv.org/abs/1812.01723

Examples

# use the simulated data provided in the package
covX = as.matrix(sim_rc[,5:8])
# Implement unnormalized IPW DID estimator
ipw_did_rc(y = sim_rc$y, post = sim_rc$post, D = sim_rc$d,
           covariates= covX)

#>  Call:
#> ipw_did_rc(y = sim_rc$y, post = sim_rc$post, D = sim_rc$d, covariates = covX)
#> ------------------------------------------------------------------
#>  IPW DID estimator for the ATT:
#>  
#>    ATT     Std. Error  t value    Pr(>|t|)  [95% Conf. Interval] 
#>  -19.8933   53.8682    -0.3693     0.7119    -125.475   85.6884  
#> ------------------------------------------------------------------
#>  Estimator based on (stationary) repeated cross-sections data.
#>  Horvitz-Thompson-type IPW estimator.
#>  Propensity score est. method: maximum likelihood.
#>  Analytical standard error.
#> ------------------------------------------------------------------
#>  See Sant'Anna and Zhao (2020) for details.