Standardized inverse probability weighted DiD estimator, with panel data

std_ipw_did_panel is used to compute inverse probability weighted (IPW) estimators for the ATT in difference-in-differences (DiD) setups with panel data. IPW weights are normalized to sum up to one, that is, the estimator is of the Hajek type.

std_ipw_did_panel(
  y1,
  y0,
  D,
  covariates,
  i.weights = NULL,
  boot = FALSE,
  boot.type = "weighted",
  nboot = NULL,
  inffunc = FALSE
)

Arguments

y1
An \(n\) x \(1\) vector of outcomes from the post-treatment period.

y0
An \(n\) x \(1\) vector of outcomes from the pre-treatment period.

D
An \(n\) x \(1\) vector of Group indicators (=1 if observation is treated in the post-treatment, =0 otherwise).

covariates
An \(n\) x \(k\) matrix of covariates to be used in the propensity score estimation. If covariates = NULL, this leads to an unconditional DID estimator.

i.weights
An \(n\) x \(1\) vector of weights to be used. If NULL, then every observation has the same weights.

boot
Logical argument to whether bootstrap should be used for inference. Default is FALSE.

boot.type
Type of bootstrap to be performed (not relevant if boot = FALSE). Options are "weighted" and "multiplier". If boot = TRUE, default is "weighted".

nboot
Number of bootstrap repetitions (not relevant if boot = FALSE). Default is 999.

inffunc
Logical argument to whether influence function should be returned. Default is FALSE.

Value

A list containing the following components:

ATT

The IPW DID point estimate.

se

The IPW DID standard error

uci

Estimate of the upper bound of a 95% CI for the ATT

lci

Estimate of the lower bound of a 95% CI for the ATT

boots

All Bootstrap draws of the ATT, in case bootstrap was used to conduct inference. Default is NULL

att.inf.func

Estimate of the influence function. Default is NULL

call.param

The matched call.

argu

Some arguments used (explicitly or not) in the call (panel = TRUE, normalized = TRUE, boot, boot.type, nboot, type="ipw")

References

Abadie, Alberto (2005), "Semiparametric Difference-in-Differences Estimators", Review of Economic Studies, vol. 72(1), p. 1-19, doi: 10.1111/0034-6527.00321 .

Sant'Anna, Pedro H. C. and Zhao, Jun. (2020), "Doubly Robust Difference-in-Differences Estimators." Journal of Econometrics, Vol. 219 (1), pp. 101-122, doi: 10.1016/j.jeconom.2020.06.003

Examples

# Form the Lalonde sample with CPS comparison group
eval_lalonde_cps <- subset(nsw, nsw$treated == 0 | nsw$sample == 2)
# Select some covariates
covX = as.matrix(cbind(eval_lalonde_cps$age, eval_lalonde_cps$educ,
                       eval_lalonde_cps$black, eval_lalonde_cps$married,
                       eval_lalonde_cps$nodegree, eval_lalonde_cps$hisp,
                       eval_lalonde_cps$re74))
# Implement normalized IPW DID with panel data
std_ipw_did_panel(y1 = eval_lalonde_cps$re78, y0 = eval_lalonde_cps$re75,
                D = eval_lalonde_cps$experimental,
                covariates = covX)

#>  Call:
#> std_ipw_did_panel(y1 = eval_lalonde_cps$re78, y0 = eval_lalonde_cps$re75, 
#>     D = eval_lalonde_cps$experimental, covariates = covX)
#> ------------------------------------------------------------------
#>  IPW DID estimator for the ATT:
#>  
#>    ATT     Std. Error  t value    Pr(>|t|)  [95% Conf. Interval] 
#> -1021.6095  397.5322   -2.5699     0.0102   -1800.7726 -242.4464 
#> ------------------------------------------------------------------
#>  Estimator based on panel data.
#>  Hajek-type IPW estimator (weights sum up to 1).
#>  Propensity score est. method: maximum likelihood.
#>  Analytical standard error.
#> ------------------------------------------------------------------
#>  See Sant'Anna and Zhao (2020) for details.