Locally Efficient Doubly Robust Difference-in-Differences Estimators for the ATT

drdid computes the doubly robust estimators for the average treatment effect on the treated in DID setups. It can be used with panel or repeated cross section data. The nuisance parameters can be estimated using "traditional" methods (OLS for the outcome regressions and maximum likelihood for the propensity score), or "improved" methods (weighted least squares for the outcome regressions and inverse probability tilting for the propensity score). Estimators based on "improved methods" are also doubly robust for inference. See Sant'Anna and Zhao (2020) for a detailed description.

drdid(
  yname,
  tname,
  idname,
  dname,
  xformla = NULL,
  data,
  panel = TRUE,
  estMethod = c("imp", "trad"),
  weightsname = NULL,
  boot = FALSE,
  boot.type = c("weighted", "multiplier"),
  nboot = 999,
  inffunc = FALSE
)

Arguments

yname
The name of the outcome variable.

tname
The name of the column containing the time periods.

idname
The name of the column containing the unit id name.

dname
The name of the column containing the treatment group (=1 if observation is treated in the post-treatment, =0 otherwise)

xformla
A formula for the covariates to include in the model. It should be of the form ~ X1 + X2 (intercept should not be listed as it is always automatically included). Default is NULL which is equivalent to xformla=~1.

data
The name of the data.frame that contains the data.

panel
Whether or not the data is a panel dataset. The panel dataset should be provided in long format -- that is, where each row corresponds to a unit observed at a particular point in time. The default is TRUE. Whens panel=TRUE, the variable idname must be set. When panel=FALSE, the data is treated as stationary repeated cross sections.

estMethod
the method to estimate the nuisance parameters. The default is "imp" which uses weighted least squares to estimate the outcome regressions and inverse probability tilting to the estimate the the propensity score, leading to the improved locally efficient DR DID estimator proposed by Sant'Anna and Zhao (2020). The other alternative is "trad", which then uses OLS to estimate outcome regressions and maximum likelihood to estimate propensity score. This leads to the "traditional" locally efficient DR DID estimator proposed by Sant'Anna and Zhao (2020).

weightsname
The name of the column containing the sampling weights. If NULL, then every observation has the same weights.

boot
Logical argument to whether bootstrap should be used for inference. Default is FALSE and analytical standard errors are reported.

boot.type
Type of bootstrap to be performed (not relevant if boot = FALSE). Options are "weighted" and "multiplier". If boot==TRUE, default is "weighted".

nboot
Number of bootstrap repetitions (not relevant if boot = FALSE). Default is 999.

inffunc
Logical argument to whether influence function should be returned. Default is FALSE.

Value

A list containing the following components:

ATT

The DR DID point estimate

se

The DR DID standard error

uci

Estimate of the upper bound of a 95% CI for the ATT

lci

Estimate of the lower bound of a 95% CI for the ATT

boots

All Bootstrap draws of the ATT, in case bootstrap was used to conduct inference. Default is NULL

att.inf.func

Estimate of the influence function. Default is NULL

ps.flag

Convergence Flag for the propensity score estimation (only active if estMethod = "imp".): =0 if trust algorithm converged, =1 if IPT (original) algorithm converged (in case it was used), =2 if GLM logit estimator was used (i.e., if both trust and IPT did not converged).

call.param

The matched call.

argu

Some arguments used in the call (panel, estMethod, boot, boot.type, nboot, type="dr")

References

Sant'Anna, Pedro H. C. and Zhao, Jun. (2020), "Doubly Robust Difference-in-Differences Estimators." Journal of Econometrics, Forthcoming, https://arxiv.org/abs/1812.01723

Examples

# -----------------------------------------------
# Panel data case
# -----------------------------------------------
# Form the Lalonde sample with CPS comparison group
eval_lalonde_cps <- subset(nsw_long, nsw_long$treated == 0 | nsw_long$sample == 2)
# -----------------------------------------------
# Implement improved DR locally efficient DID with panel data
drdid(yname="re", tname = "year", idname = "id", dname = "experimental",
      xformla= ~ age+ educ+ black+ married+ nodegree+ hisp+ re74,
      data = eval_lalonde_cps, panel = TRUE)

#> Error in pre_process_drdid(yname = yname, tname = tname, idname = idname,     dname = dname, xformla = xformla, data = data, panel = panel,     estMethod = estMethod, weightsname = weightsname, boot = boot,     boot.type = boot.type, nboot = nboot, inffunc = inffunc): dname = experimental could not be found in the data provided.

#Implement "traditional" DR locally efficient DID with panel data
drdid(yname="re", tname = "year", idname = "id", dname = "experimental",
      xformla= ~ age+ educ+ black+ married+ nodegree+ hisp+ re74,
      data = eval_lalonde_cps, panel = TRUE, estMethod = "trad")

#> Error in pre_process_drdid(yname = yname, tname = tname, idname = idname,     dname = dname, xformla = xformla, data = data, panel = panel,     estMethod = estMethod, weightsname = weightsname, boot = boot,     boot.type = boot.type, nboot = nboot, inffunc = inffunc): dname = experimental could not be found in the data provided.

# -----------------------------------------------
# Repeated cross section case
# -----------------------------------------------
# use the simulated data provided in the package
#Implement "improved" DR locally efficient DID with repeated cross-section data
drdid(yname="y", tname = "post", idname = "id", dname = "d",
      xformla= ~ x1 + x2 + x3 + x4,
      data = sim_rc, panel = FALSE, estMethod = "imp")

#>  Call:
#> drdid(yname = "y", tname = "post", idname = "id", dname = "d", 
#>     xformla = ~x1 + x2 + x3 + x4, data = sim_rc, panel = FALSE, 
#>     estMethod = "imp")
#> ------------------------------------------------------------------
#>  Further improved locally efficient DR DID estimator for the ATT:
#>  
#>    ATT     Std. Error  t value    Pr(>|t|)  [95% Conf. Interval] 
#>  -0.2089     0.2003    -1.0425     0.2972    -0.6015     0.1838  
#> ------------------------------------------------------------------
#>  Estimator based on (stationary) repeated cross-sections data.
#>  Outcome regression est. method: weighted least squares.
#>  Propensity score est. method: inverse prob. tilting.
#>  Analytical standard error.
#> ------------------------------------------------------------------
#>  See Sant'Anna and Zhao (2020) for details.

#Implement "traditional" DR locally efficient DID with repeated cross-section data
drdid(yname="y", tname = "post", idname = "id", dname = "d",
      xformla= ~ x1 + x2 + x3 + x4,
      data = sim_rc, panel = FALSE, estMethod = "trad")

#>  Call:
#> drdid(yname = "y", tname = "post", idname = "id", dname = "d", 
#>     xformla = ~x1 + x2 + x3 + x4, data = sim_rc, panel = FALSE, 
#>     estMethod = "trad")
#> ------------------------------------------------------------------
#>  Locally efficient DR DID estimator for the ATT:
#>  
#>    ATT     Std. Error  t value    Pr(>|t|)  [95% Conf. Interval] 
#>  -0.1678     0.2009    -0.8352     0.4036    -0.5616     0.226   
#> ------------------------------------------------------------------
#>  Estimator based on (stationary) repeated cross-sections data.
#>  Outcome regression est. method: OLS.
#>  Propensity score est. method: maximum likelihood.
#>  Analytical standard error.
#> ------------------------------------------------------------------
#>  See Sant'Anna and Zhao (2020) for details.